housing-tenure/requirements.txt
# Housing Tenure Analysis
geopandas>=0.14.0
pandas>=2.0.0
numpy>=1.24.0
census>=0.8.21
requests>=2.31.0
python-dotenv>=1.0.0
scipy>=1.11.0
